allow zero buffer for zaptec blueprint
Nice job, tested it a bit and it seems to work great :)

One of my phases are heavily used now during the winter sometimes I might have exactly 6A available on that phase for a long time, so I would like to be able to charge during this time. My fuses can withstand a bit of overconsumption a few minutes so no real problem if it goes over a little bit for ~20 seconds.